pyproject.toml
[tool.poetry]
name = "matlab-ai-agent"
version = "0.1.0"
description = "AI agent for generating and executing MATLAB simulations"
authors = ["Marco Melloni <marcomelloni01@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/yourusername/matlab-ai-agent"
documentation = "https://github.com/yourusername/matlab-ai-agent/tree/main/docs"
keywords = ["matlab", "ai", "agent", "simulation", "ode"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
packages = [
    { include = "src"}
]

[tool.poetry.dependencies]
python = "^3.8"
openai = "^1.3.0"
python-dotenv = "^1.0.0"
click = "^8.1.3"
colorlog = "^6.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
pylint = "^2.17.4"
pre-commit = "^3.3.2"
autopep8 = "^2.0.2"
coverage = "^7.2.7"
pytest-cov = "^4.1.0"
tox = "^4.6.0"

[tool.poetry.scripts]
matlab-ai-agent = "src.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.pylint.messages_control]
disable = [
    "C0103", # variable-name
    "C0111", # missing-docstring
    "R0903", # too-few-public-methods
    "R0913", # too-many-arguments
]

[tool.pylint.format]
max-line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=src --cov-report=term-missing --cov-report=xml"